#125 move new color back to white: enoki (`#FAFBEF`)
